Move SpecVersion in field_types module
