+++ cloudevents_pydantic/events/field_types/__init__.py
@@ -27,6 +27,7 @@ from ._canonic_types import (
     Boolean,
     DateTime,
     Integer,
+    SpecVersion,
     String,
     URIReference,
 )
